+++ src/main/java/com/aakivaa/emss/models/EventsCostRate.java
@@ -1,5 +1,6 @@
 package com.aakivaa.emss.models;
 
+import com.aakivaa.emss.models.users.Venue;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 
@@ -17,7 +18,7 @@ public class EventsCostRate implements Serializable {
     @Id
     @SequenceGenerator(name = "function_id_sequence", sequenceName = "function_id_sequence")
     @GeneratedValue(generator = "function_id_sequence", strategy = GenerationType.SEQUENCE)
-    private Integer id;
+    private Long id;
 
     private Double marriageCost;
