+++ src/main/java/com/aakivaa/emss/repo/FunctionRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FunctionRepo extends JpaRepository<EventsCostRate, Integer> {
+public interface FunctionRepo extends JpaRepository<EventsCostRate, Long> {
 
 }
